[ebpf backend] update the semantics of calls
The compcert stack pointer is mapped to the shadow BP register.
The eBPF R10 (SP) register is BP+ofs which is the top of
the stack frame (excluding return address and old stack pointer).
In the code, R10 is never modified.
The creation of stackframes is modelled by the pseudo-instructions
allocframe and freeframe which do not generate code.

callee-saved are explicitly spill and restored, though this might also
be performed by the call and ret instructions.